requirements.txt
asgiref==3.4.1
Django==3.2.7
image==1.5.33
numpy==1.21.2
Pillow==8.3.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
